trocando os metodos de pagamento (cadastro,atualizar,consultar) para sem o fxml

+++ src/main/java/br/pucpr/omcejavafx/Pagamento/Pagamento.java
@@ -5,11 +5,11 @@ import java.io.Serializable;
 public class Pagamento implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    private long id;
+    private int id;
     private String metodoPagamento;
     private String data;
 
-    public Pagamento(long id, String metodoPagamento, String data) {
+    public Pagamento(int id, String metodoPagamento, String data) {
         this.id = id;
         this.metodoPagamento = metodoPagamento;
         this.data = data;
@@ -18,8 +18,8 @@ public class Pagamento implements Serializable {
     public Pagamento() {
     }
 
-    public long getId() { return id; }
-    public void setId(long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
 
 
     public String getMetodoPagamento() { return metodoPagamento; }
